move APIFnDebug to action_debug.go
